requirements.txt
streamlit
pandas
numpy
pulp
scipy
plotly
supabase
google-generativeai